rename kd.c->nco_kd.c kd.h->nco_kd.h
