+++ src/nco/nco_kd.h
